requirements.txt
alabaster==0.7.12
alembic==1.8.1
awscli==1.26.4
Babel==2.10.3
botocore==1.28.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.4
contourpy==1.0.5
coverage==6.5.0
cycler==0.11.0
databricks-cli==0.17.3
docker==6.0.1
docutils==0.16
entrypoints==0.4
flake8==5.0.4
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
greenlet==2.0.0
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-dataclass==8.5.9
matplotlib==3.6.0
mccabe==0.7.0
mlflow==1.30.0
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
protobuf==4.21.9
pyasn1==0.4.8
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.42
sqlparse==0.4.3
sweetviz==2.1.4
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
websocket-client==1.4.1
Werkzeug==2.2.2
zipp==3.10.0
